Refactor: 주소 가져오는 코드 커스텀훅 화
