+++ src/constants/const.js
@@ -4,3 +4,9 @@ export const CATEGORY = {
   TIMETABLE: 'TIMETABLE',
   MYINFO: 'MYINFO',
 };
+
+export const geolocationOptions = {
+  enableHighAccuracy: true,
+  timeout: 1000 * 60 * 1, // 1 min (1000 ms * 60 sec * 1 minute = 60 000ms)
+  maximumAge: 1000 * 3600 * 24, // 24 hour
+};
